updates to fixtures
