+++ tests/fixtures/metatranscriptome_assembly_set.json
@@ -1,10 +1,10 @@
 [
   {
     "has_input": [
-      "nmdc:dobj-11-filteredreads"
+      "nmdc:dobj-11-filteredreads1"
     ],
     "part_of": [
-      "nmdc:omprc-11-metag1"
+      "nmdc:omprc-11-metat1"
     ],
     "scaf_N50": 302542,
     "ctg_logsum": 5909271,
@@ -14,26 +14,26 @@
     "scaf_logsum": 5934400,
     "scaf_pct_gt50K": 1.7355337,
     "gap_pct": 0.00334,
-    "git_url": "https://github.com/microbiomedata/metaAssembly",
-    "version": "v1.0.3",
+    "git_url": "https://github.com/microbiomedata/metaT_Assembly",
+    "version": "v0.0.1",
     "has_output": [
-      "nmdc:dobj-11-contigs",
-      "nmdc:dobj-11-coverstats",
-      "nmdc:dobj-11-coverbam",
-      "nmdc:dobj-11-assemblyagp",
-      "nmdc:dobj-11-scaffolds"
+      "nmdc:dobj-11-contigst1",
+      "nmdc:dobj-11-coverstatst1",
+      "nmdc:dobj-11-coverbamt1",
+      "nmdc:dobj-11-assemblyinfot1",
+      "nmdc:dobj-11-bamidxt1"
     ],
     "asm_score": 11.297,
-    "was_informed_by": "nmdc:omprc-11-metag1",
+    "was_informed_by": "nmdc:omprc-11-metat1",
     "ctg_powsum": 727370,
     "scaf_max": 517431,
-    "id": "nmdc:wfmgas-11-test001.1",
+    "id": "nmdc:wfmtas-11-test001.1",
     "scaf_powsum": 730816,
     "execution_resource": "NERSC-Cori",
     "contigs": 1705758,
     "scaf_n_gt50K": 216,
     "ctg_N50": 304732,
-    "name": "Assembly Activity for nmdc:mga0vx38",
+    "name": "Assembly Activity for metaT",
     "ctg_max": 517431,
     "gc_std": 0.06928,
     "contig_bp": 1168572354,
@@ -43,7 +43,7 @@
     "started_at_time": "2021-08-05T14:48:51+00:00",
     "ctg_L90": 322,
     "scaf_bp": 1168611354,
-    "type": "nmdc:MetagenomeAssembly",
+    "type": "nmdc:MetatranscriptomeAssembly",
     "scaf_L50": 789,
     "scaffolds": 1702137,
     "ended_at_time": "2021-09-15T10:13:20+00:00"
