+++ tests/fixtures/metatranscriptome_expression_analysis_set.json
@@ -1,24 +1,24 @@
 [
   {
-    "id": "nmdc:wfmtex-11-testmetag1.1",
-    "name": "Metatranscriptome Expression Analysis for nmdc:wfrqc-11-test-metat1.1",
+    "id": "nmdc:wfmtex-11-metat1.1",
+    "name": "Metatranscriptome Expression Analysis for nmdc:wfmtex-11-metat1.1",
     "started_at_time": "2024-04-11T20:51:15.535533+00:00",
     "ended_at_time": "2024-04-11T23:39:28.659534+00:00",
-    "was_informed_by": "nmdc:omprc-11-testmetat1",
+    "was_informed_by": "nmdc:omprc-11-metat1",
     "execution_resource": "NERSC-Perlmutter",
     "git_url": "https://github.com/microbiomedata/metaT_ReadCounts",
     "has_input": [
       "nmdc:dobj-11-functionalt1",
-      "nmdc:dobj-11-rawreads2"
+      "nmdc:dobj-11-contigmappingt1",
+      "nmdc:dobj-11-coverbamt1"
     ],
-    "type": "nmdc:ReadQcAnalysisActivity",
+    "type": "nmdc:MetatranscriptomeExpressionAnalysis",
     "has_output": [
-      "nmdc:dobj-11-qcstats",
-      "nmdc:dobj-11-qcinfo",
-      "nmdc:dobj-11-filteredreads"
+      "nmdc:dobj-11-expcountst1",
+      "nmdc:dobj-11-expinfot1"
     ],
     "part_of": [
-      "nmdc:omprc-11-test-metag1"
+      "nmdc:omprc-11-metat1"
     ],
     "version": "v1.0.8"
   }
